Refactor client accept into client module
